requirements.txt
# PDF Processing
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow==10.1.0

# Google Sheets API
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Data Processing
pandas==2.1.4

# Utilities
python-dotenv==1.0.0
